+++ examples/App.js
@@ -1,7 +1,7 @@
-import React, {PureComponent} from 'react';
-import {render} from 'react-dom';
+import React, { PureComponent } from 'react';
+import { render } from 'react-dom';
 
-import '../src/billboard.css';
+import 'billboard.js/dist/billboard.css';
 
 // charts
 import BarChart from './BarChart';
@@ -11,14 +11,26 @@ import LineChart from './LineChart';
 class App extends PureComponent {
   element = null;
 
+  state = { visible: false };
+
+  componentDidMount() {
+    setTimeout(() => {
+      this.setState(() => ({ visible: true }));
+    }, 0);
+  }
+
   render() {
     return (
       <div>
         <h1>App</h1>
 
-        <LineChart />
-        <BarChart />
-        <DonutChart />
+        {this.state.visible && (
+          <div>
+            <LineChart />
+            <BarChart />
+            <DonutChart />
+          </div>
+        )}
       </div>
     );
   }